refactor: move processCompileAction into its own file.
test: Capture all console output during tests.
