+++ packages/cspell-tools/src/app.ts
@@ -1,17 +1,15 @@
 // For large dictionaries, it is necessary to increase the memory limit.
 
-import { compileWordList, compileTrie, Logger } from './compiler';
+import { Logger } from './compiler';
 import * as compiler from './compiler';
 import * as path from 'path';
 import * as program from 'commander';
 import glob from 'glob';
-import { genSequence, Sequence } from 'gensequence';
-import { streamWordsFromFile } from './compiler/iterateWordsFromFile';
-import { ReaderOptions } from './compiler/Reader';
+import { processCompileAction } from './processCompileAction';
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const npmPackage = require(path.join(__dirname, '..', 'package.json'));
 
-function globP(pattern: string): Promise<string[]> {
+export function globP(pattern: string): Promise<string[]> {
     // Convert windows separators.
     pattern = pattern.replace(/\\/g, '/');
     return new Promise((resolve, reject) => {
@@ -21,7 +19,7 @@ function globP(pattern: string): Promise<string[]> {
     });
 }
 
-interface CompileCommonOptions {
+export interface CompileCommonOptions {
     output?: string;
     compress: boolean;
     max_depth?: string;
@@ -48,7 +46,7 @@ interface CompileTrieOptions extends CompileCommonOptions {
 }
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
-const log: Logger = (message?: any, ...optionalParams: any[]) => {
+export const log: Logger = (message?: any, ...optionalParams: any[]) => {
     console.log(`${new Date().toISOString()} ${message}`, ...optionalParams);
 };
 
@@ -94,7 +92,7 @@ export function run(program: program.Command, argv: string[]): Promise<void> {
             .option('--trie', 'Compile into a trie file.', false)
             .option('--no-sort', 'Do not sort the result')
             .action((src: string[], options: CompileOptions) => {
-                const result = processAction(src, options);
+                const result = processCompileAction(src, options);
                 resolve(result);
             });
 
@@ -105,7 +103,7 @@ export function run(program: program.Command, argv: string[]): Promise<void> {
                     'Compile words lists or Hunspell dictionary into trie files used by cspell.\nAlias of `compile --trie`'
                 )
         ).action((src: string[], options: CompileTrieOptions) => {
-            const result = processAction(src, { ...options, trie: true });
+            const result = processCompileAction(src, { ...options, trie: true });
             resolve(result);
         });
 
@@ -121,127 +119,3 @@ export function run(program: program.Command, argv: string[]): Promise<void> {
         resolve();
     });
 }
-
-interface FileToProcess {
-    src: string;
-    words: Sequence<string>;
-}
-
-function parseNumber(s: string | undefined): number | undefined {
-    const n = parseInt(s ?? '');
-    return isNaN(n) ? undefined : n;
-}
-
-type ActionFn = (words: Sequence<string>, dst: string) => Promise<unknown>;
-
-async function processAction(src: string[], options: CompileCommonOptions): Promise<void> {
-    const useTrie = options.trie || options.trie3 || options.trie4 || false;
-    const fileExt = useTrie ? '.trie' : '.txt';
-    console.log(
-        'Compile:\n output: %s\n compress: %s\n files:\n  %s \n\n',
-        options.output || 'default',
-        options.compress ? 'true' : 'false',
-        src.join('\n  ')
-    );
-    const experimental = new Set(options.experimental);
-    const skipNormalization = experimental.has('compound');
-    const { keepRawCase = false, split: splitWords = false, sort = true, useLegacySplitter: legacy } = options;
-
-    const action = useTrie
-        ? async (words: Sequence<string>, dst: string) => {
-              return compileTrie(words, dst, {
-                  ...options,
-                  skipNormalization,
-                  splitWords,
-                  keepRawCase,
-                  legacy,
-                  base: parseNumber(options.trieBase),
-                  sort: false,
-              });
-          }
-        : async (src: Sequence<string>, dst: string) => {
-              return compileWordList(src, dst, {
-                  splitWords,
-                  sort,
-                  skipNormalization,
-                  keepRawCase,
-                  legacy,
-              }).then(() => src);
-          };
-    const ext = fileExt + (options.compress ? '.gz' : '');
-    const maxDepth = parseNumber(options.max_depth);
-    const useAnnotation = experimental.has('compound');
-    const readerOptions: ReaderOptions = { maxDepth, useAnnotation };
-
-    const globResults = await Promise.all(src.map((s) => globP(s)));
-    const filesToProcess = genSequence(globResults)
-        .concatMap((files) => files)
-        .map(async (filename) => {
-            log(`Reading ${path.basename(filename)}`);
-            const words = await streamWordsFromFile(filename, readerOptions);
-            log(`Done reading ${path.basename(filename)}`);
-            const f: FileToProcess = {
-                src: filename,
-                words,
-            };
-            return f;
-        });
-
-    const r = options.merge
-        ? processFiles(action, filesToProcess, toMergeTargetFile(options.merge, options.output, ext))
-        : processFilesIndividually(action, filesToProcess, (s) => toTargetFile(s, options.output, ext));
-    await r;
-    log(`Complete.`);
-}
-
-function toFilename(name: string, ext: string) {
-    return path.basename(name).replace(/((\.txt|\.dic|\.aff|\.trie)(\.gz)?)?$/, '') + ext;
-}
-
-function toTargetFile(filename: string, destination: string | undefined, ext: string) {
-    const outFileName = toFilename(filename, ext);
-    const dir = destination ?? path.dirname(filename);
-    return path.join(dir, outFileName);
-}
-
-function toMergeTargetFile(filename: string, destination: string | undefined, ext: string) {
-    const outFileName = path.join(path.dirname(filename), toFilename(filename, ext));
-    return path.resolve(destination ?? '.', outFileName);
-}
-
-async function processFilesIndividually(
-    action: ActionFn,
-    filesToProcess: Sequence<Promise<FileToProcess>>,
-    srcToTarget: (_src: string) => string
-) {
-    const toProcess = filesToProcess.map(async (pFtp) => {
-        const { src, words } = await pFtp;
-        const dst = srcToTarget(src);
-        log('Process "%s" to "%s"', src, dst);
-        await action(words, dst);
-        log('Done "%s" to "%s"', src, dst);
-    });
-
-    for (const p of toProcess) {
-        await p;
-    }
-}
-
-async function processFiles(action: ActionFn, filesToProcess: Sequence<Promise<FileToProcess>>, mergeTarget: string) {
-    const toProcess = await Promise.all([...filesToProcess]);
-    const dst = mergeTarget;
-
-    const words = genSequence(toProcess)
-        .map((ftp) => {
-            const { src } = ftp;
-            log('Process "%s" to "%s"', src, dst);
-            return ftp;
-        })
-        .concatMap((ftp) => ftp.words);
-    await action(words, dst);
-    log('Done "%s"', dst);
-}
-
-export const __testing__ = {
-    processAction,
-};
